refactor(kibisis): move wallet to signle file structure
