+++ packages/use-wallet/src/wallets/index.ts
@@ -2,6 +2,7 @@ export * from './base'
 export * from './custom'
 export * from './defly'
 export * from './exodus'
+export * from './kibisis'
 export * from './kmd'
 export * from './magic'
 export * from './mnemonic'
