--- packages/use-wallet/src/wallets/kibisis/index.ts
@@ -1,3 +1,0 @@
-export { default as KibisisWallet } from './client'
-export * from './constants'
-export * from './utils'
